Experimental uint128_t support (#248)
* Experimental uint128_t support

* ULLL not available

* ONE_BCI safety

* ONE_BCI safety

* Type safety (ONE_BCI)

* Triggering CI

* Per review

* Per review

+++ examples/quantum_associative_memory.cpp
@@ -64,6 +64,6 @@ int main()
             outputLayer[i]->Predict();
         }
         comp = qReg->MReg(InputCount, OutputCount);
-        std::cout << "Input: " << (int)perm << ", Output: " << (int)comp << std::endl;
+        std::cout << "Input: " << perm << ", Output: " << comp << std::endl;
     }
 }
